python/requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
grpcio==1.31.0
grpcio-tools==1.31.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
protobuf==3.12.4
pycodestyle==2.6.0
pylint==2.5.3
pymilvus==0.2.14
requests==2.24.0
six==1.15.0
toml==0.10.1
ujson==3.1.0
urllib3==1.25.10
wrapt==1.12.1
